pyproject.toml
[project]
name = "hitcoursehunter"
version = "0.1.0"
description = "A course selection assistant tool for HITSZ"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "flask>=2.3.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selectolax>=0.3.31",
]

[dependency-groups]
dev = ["pre-commit>=4.2.0", "pyright>=1.1.402", "ruff>=0.12.0"]
